third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/customizable-select/selectedoption.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] The <selectedoption> element should reflect the HTML contents of the selected <option>.
  assert_equals: The innerHTML of <selectedoption> should initially match the innerHTML of the selected <option>. expected "\\n      <span class=\\"one\\">span</span> one\\n    " but got ""
Harness: the test ran to completion.
